Attempt to make pure python branch merge.
Reverted most changes in algos3d and humanmodifier since rev 3593, except for those necessary to make warping work. The human is accessed through a global variable, which is not so pretty but works. The isWarp attribute is accessed through a hasattr. Hopefully the changes to human.py will be acceptable.

git-svn-id: http://makehuman.googlecode.com/svn@3629 ea038faa-237f-11de-8bc4-bb72633a7edc